refactor: Reorganize debug tools and improve service architecture
- Move debug_window_structure.py to debugs/ directory for better organization
- Replace official_collector.py with official_service.py for clearer naming
- Replace command_executor.py with command_service.py for consistency
- Update CLAUDE.md with comprehensive debugging documentation
- Enhance dialog_service.py with improved error handling
- Update main.py imports to reflect new service structure
