+++ debugs/debug_window_structure.py
